+++ mix/test_cldr.ex
@@ -7,6 +7,5 @@ defmodule Money.Cldr do
   use Cldr,
     locales: ["en", "de", "it", "es", "fr", "da", "zh-Hant-HK", "zh-Hans", "ja"],
     default_locale: "en",
-    data_dir: "../cldr/priv/cldr",
     providers: [Cldr.Number, Money]
 end